JDBC add new repositories

+++ src/main/java/ua/kobzev/theatre/repository/impl/inmemory/UserRepositoryImpl.java
@@ -42,8 +42,8 @@ public class UserRepositoryImpl implements UserRepository {
 	}
 
 	@Override
-	public User getById(int id) {
-		return usersList.stream().filter(user -> ((Integer) user.getId()).equals(id)).findFirst().orElse(null);
+	public User getById(Integer id) {
+		return usersList.stream().filter(user -> user.getId().equals(id)).findFirst().orElse(null);
 	}
 
 	@Override
